+++ tests/test_threadpoolctl.py
@@ -11,8 +11,9 @@ import shutil
 
 import threadpoolctl
 from threadpoolctl import threadpool_limits, threadpool_info
-from threadpoolctl import ThreadpoolController
+from threadpoolctl import LibController, ThreadpoolController
 from threadpoolctl import _ALL_PREFIXES, _ALL_USER_APIS
+from threadpoolctl import _determine_thread_limit_scope
 
 from .utils import cython_extensions_compiled
 from .utils import check_nested_prange_blas
@@ -620,6 +621,7 @@ def test_architecture():
     expected_openblas_architectures = (
         # XXX: add more as needed by CI or developer laptops
         "armv8",
+        "barcelona",
         "cooperlake",
         "haswell",
         "neoversen1",
@@ -905,3 +907,118 @@ def test_windows_library_path_exceeds_internal_limit(tmp_path, monkeypatch):
                 filepath = "\\\\" + filepath[4:]
         filepaths.add(os.path.normcase(os.path.normpath(filepath)))
     assert expected_path not in filepaths
+
+
+def parse_version(version: str) -> list[int]:
+    return list(map(int, version.split(".")))
+
+
+@pytest.fixture(
+    params=[
+        (
+            {"internal_api": "openblas"},
+            lambda lib: (
+                lib.threading_layer == "openmp"
+                # For Windows support, see
+                # https://github.com/joblib/threadpoolctl/issues/230
+                and sys.platform in ("linux", "darwin")
+                and parse_version(lib.version) >= parse_version("0.3.34")
+            ),
+        ),
+        (
+            {"internal_api": "mkl"},
+            lambda _lib: True,
+        ),
+    ],
+    # ids correspond to the params above:
+    ids=["openblas-openmp", "mkl"],
+)
+def thread_local_blas_lib(request) -> LibController:
+    """Create all LibControllers that use a thread-local setting."""
+    select_filter, extra_check = request.param
+    controller = ThreadpoolController().select(**select_filter)
+    if not controller.lib_controllers:
+        pytest.skip(f"{select_filter} controller not found")
+
+    libs = [
+        lib
+        for lib in controller.lib_controllers
+        if extra_check(lib) and lib.internal_api == select_filter["internal_api"]
+    ]
+    if not libs:
+        pytest.skip("No libraries matched the requirements")
+
+    assert len(libs) == 1
+    return libs[0]
+
+
+def test_setting_limit_on_thread_local_blas_api_is_reported_as_thread_local(
+    thread_local_blas_lib: LibController,
+) -> None:
+    """
+    Setting the number of threads for libraries that support thread-local
+    setting API is reported as doing so.
+
+    This doesn't check actual behavior, only reported behavior.
+    """
+    lib = thread_local_blas_lib
+    scope = _determine_thread_limit_scope(lib.get_num_threads, lib.set_num_threads)
+    assert scope == "current_thread"
+
+
+@pytest.mark.skipif(
+    sys.platform != "linux" or shutil.which("strace") is None,
+    reason="requires strace on Linux",
+)
+def test_setting_limit_on_thread_local_blas_api_is_actually_thread_local(
+    thread_local_blas_lib: LibController,
+) -> None:
+    """
+    Setting the number of threads for libraries that support thread-local
+    setting API actually does so.
+    """
+
+    # The test script uses NumPy, there might be multiple BLAS in this test
+    # process, and we want to only run if _NumPy_ uses that library.
+    # So check that before proceeding.
+    output = json.loads(
+        subprocess.check_output(
+            [
+                "python",
+                "-c",
+                "import numpy, json, threadpoolctl; print(json.dumps(threadpoolctl.threadpool_info()))",
+            ]
+        )
+    )
+    found_correct_blas = False
+    for library in output:
+        if library["internal_api"] == thread_local_blas_lib.internal_api:
+            found_correct_blas = True
+            break
+    if not found_correct_blas:
+        pytest.skip("NumPy doesn't use the BLAS we want to test")
+
+    def num_threads_created(limit: int) -> int:
+        result = 0
+        for line in subprocess.check_output(
+            [
+                "strace",
+                "-f",
+                "-e",
+                "clone3",
+                "python",
+                "-m",
+                "tests._limit_blas",
+                str(limit),
+            ],
+            stderr=subprocess.STDOUT,
+        ).splitlines():
+            if b"clone3(" in line and b"CLONE_THREAD" in line:
+                result += 1
+        return result
+
+    # _limit_blas runs BLAS operations in 2 Python threads, so by changing the
+    # BLAS limit from 1 to 4 we expect an extra 2 * (4 - 1) == 6 threads.
+    nmc_1 = num_threads_created(1)
+    nmc_4 = num_threads_created(4)
+    assert nmc_4 - nmc_1 == 6